Move `Observable::fromPromise()` to `React\Promise::toObservable()`
