+++ lib/Rx/Operator/DeferOperator.php
@@ -8,6 +8,7 @@ use React\Promise\PromiseInterface;
 use Rx\Observable\BaseObservable;
 use Rx\ObservableInterface;
 use Rx\ObserverInterface;
+use Rx\React\Promise;
 use Rx\SchedulerInterface;
 
 class DeferOperator implements OperatorInterface
@@ -37,7 +38,7 @@ class DeferOperator implements OperatorInterface
             $result = $factory();
 
             if ($result instanceof PromiseInterface) {
-                $result = BaseObservable::fromPromise($result);
+                $result = Promise::toObservable($result);
             }
 
             return $result->subscribe($observer);
